Abstract out `<Actions>` into a file
